Adjusted the sklearn.__version__ detection to look for version 0.20.0 instead of 0.19.0 to reflect the ve rsion difference allowing the usage of y=None in _transform_one.
Updated testing results to use float values in the 'a_GRP' instead of int to correctly reflect the bins returned when using pd.cut. The float values returned included a NaN in the results, which forces numpy to convert the array to float. All tests that previously failed due to float values being returned now pass.